test: separate TOP1M refresh matrix
